Fixed High Precision Counter upperCount being 32bit instead of 64 Rewrote ErrorMessage to use a linked list of errors for wrapping Added additional AddTask function to Scheduler

+++ inc/high_precision_counter.hpp
@@ -18,7 +18,7 @@ class HighPrecisionCounter
   private:
 	TIM_TypeDef* const tim;
 	const uint32_t timerPrecision;
-	uint32_t upperCount;
+	uint64_t upperCount;
 
 	struct DelayedCallback
 	{
